requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.6
django-cors-headers==4.9.0
django-redis==6.0.0
django-websockets==0.0.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
faiss-cpu==1.12.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.3.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-huggingface==1.2.0
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langsmith==0.5.2
MarkupSafe==3.0.2
marshmallow==3.26.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
orjson==3.11.5
packaging==25.0
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
upstash-redis==1.5.0
urllib3==2.5.0
uuid_utils==0.12.0
websockets==11.0.1
whitenoise==6.11.0
yarl==1.22.0
youtube-transcript-api==1.2.3
zope.interface==8.0.1
zstandard==0.25.0
